+++ properties/Connection/AllowToStartTwoQueriesInParallel.php
@@ -41,7 +41,7 @@ final class AllowToStartTwoQueriesInParallel implements Property
                 ->madeOf(Set::strings()->chars()->ascii())
                 ->between(0, 125),
             Set::integers(),
-        )->toSet();
+        );
     }
 
     public function applicableTo(object $connection): bool