+++ properties/Connection/CreateTableIfNotExists.php
@@ -37,7 +37,7 @@ final class CreateTableIfNotExists implements Property
             static fn(...$args) => new self(...$args),
             Name::any(),
             Column::list(),
-        )->toSet();
+        );
     }
 
     public function applicableTo(object $connection): bool