+++ properties/Connection/CreatingSameTableTwiceMustThrow.php
@@ -38,7 +38,7 @@ final class CreatingSameTableTwiceMustThrow implements Property
             static fn(...$args) => new self(...$args),
             Name::any(),
             Column::list(),
-        )->toSet();
+        );
     }
 
     public function applicableTo(object $connection): bool