+++ properties/Connection/SelectAliasedColumns.php
@@ -42,7 +42,7 @@ final class SelectAliasedColumns implements Property
                 ->madeOf(Set::strings()->chars()->ascii())
                 ->between(0, 255),
             Set::integers(),
-        )->toSet();
+        );
     }
 
     public function applicableTo(object $connection): bool