+++ properties/Connection/SelectWhere.php
@@ -45,7 +45,7 @@ final class SelectWhere implements Property
                 ->madeOf(Set::strings()->chars()->ascii())
                 ->between(0, 255),
             Set::integers(),
-        )->toSet();
+        );
     }
 
     public function applicableTo(object $connection): bool